+++ spec/graphql/mutations/team_invite/create_invite_spec.rb
@@ -16,6 +16,15 @@ RSpec.describe Mutations::TeamInvite::CreateInvite do
     res
   end
 
+  let(:result_1) do
+    res = KudoOMaticSchema.execute(
+      mutation_string,
+      context: context,
+      variables: variables
+    )
+    res
+  end
+
   let(:mutation_string) do
     %( mutation { createTeamInvite(
       teamId: "#{variables[:team_id]}"
@@ -42,6 +51,13 @@ RSpec.describe Mutations::TeamInvite::CreateInvite do
       it 'returns no errors' do
         expect(result['data']['createTeamInvite']['errors']).to be_empty
       end
+
+      it 'returns error if email addresses are already invited to team' do
+        result
+        expect(result_1['errors'].first['message']).to eq(
+          'Email test@example.com is already invited, Email test_2@example.com is already invited'
+        )
+      end
     end
 
     describe 'user is team admin' do
@@ -59,6 +75,13 @@ RSpec.describe Mutations::TeamInvite::CreateInvite do
 
       it 'returns no errors' do
         expect(result['data']['createTeamInvite']['errors']).to be_empty
+      end
+
+      it 'returns error if email addresses are already invited to team' do
+        result
+        expect(result_1['errors'].first['message']).to eq(
+          'Email test@example.com is already invited, Email test_2@example.com is already invited'
+        )
       end
     end