+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.1.1]
+### Added
+- Clear error message when team invite already exists
+
+### Fixed
+- Raise error when team invite already exists
+
 ## [4.1.0]
 ### Added
 - Check if there is at least 1 team admin per team on team member update or destroy